Move pack_circumradius to coeffs
