+++ cpp/coefficients.cpp
@@ -5,6 +5,7 @@
 // SPDX-License-Identifier:    MIT
 
 #include "coefficients.h"
+#include "geometric_quantities.h"
 #include <basix/quadrature.h>
 #include <dolfinx_cuas/QuadratureRule.hpp>
 #include <xtensor/xslice.hpp>
@@ -266,4 +267,119 @@ dolfinx_contact::pack_coefficient_quadrature(
     }
   }
   return {std::move(coefficients), cstride};
+}
+
+std::pair<std::vector<PetscScalar>, int>
+dolfinx_contact::pack_circumradius_facet(
+    std::shared_ptr<const dolfinx::mesh::Mesh> mesh,
+    const xtl::span<const std::int32_t>& active_facets)
+{
+  const int tdim = mesh->topology().dim();
+  const int gdim = mesh->geometry().dim();
+  const int fdim = tdim - 1;
+  const std::size_t num_facets = active_facets.size();
+
+  // Connectivity to evaluate at quadrature points
+  // FIXME: Move create_connectivity out of this function and call before
+  // calling the function...
+  mesh->topology_mutable().create_connectivity(fdim, tdim);
+  auto f_to_c = mesh->topology().connectivity(fdim, tdim);
+  mesh->topology_mutable().create_connectivity(tdim, fdim);
+  auto c_to_f = mesh->topology().connectivity(tdim, fdim);
+
+  // Tabulate element at quadrature points
+  // NOTE: Assuming no derivatives for now, should be reconsidered later
+  const dolfinx::mesh::CellType cell_type = mesh->topology().cell_type();
+
+  // Quadrature points for piecewise constant
+  dolfinx_cuas::QuadratureRule q_rule(cell_type, 0, fdim);
+
+  // FIXME: This does not work for prism elements
+  const std::vector<double> weights = q_rule.weights()[0];
+  const std::vector<xt::xarray<double>> points = q_rule.points();
+
+  const std::size_t num_points = weights.size();
+  const std::size_t num_local_facets = points.size();
+
+  // Get geometry data
+  const dolfinx::graph::AdjacencyList<std::int32_t>& x_dofmap
+      = mesh->geometry().dofmap();
+
+  // FIXME: Add proper interface for num coordinate dofs
+  const std::size_t num_dofs_g = x_dofmap.num_links(0);
+  xtl::span<const double> x_g = mesh->geometry().x();
+
+  // Prepare geometry data structures
+  xt::xtensor<double, 3> J = xt::zeros<double>(
+      {std::size_t(1), (std::size_t)gdim, (std::size_t)tdim});
+  xt::xtensor<double, 1> detJ = xt::zeros<double>({std::size_t(1)});
+  xt::xtensor<double, 2> coordinate_dofs
+      = xt::zeros<double>({num_dofs_g, (std::size_t)gdim});
+
+  // Get coordinate map
+  const dolfinx::fem::CoordinateElement& cmap = mesh->geometry().cmap();
+
+  xt::xtensor<double, 5> dphi_c(
+      {num_local_facets, (std::size_t)tdim, num_points, num_dofs_g, 1});
+  for (std::size_t i = 0; i < num_local_facets; i++)
+  {
+    const xt::xarray<double>& q_facet = points[i];
+    xt::xtensor<double, 4> cmap_basis_functions = cmap.tabulate(1, q_facet);
+    auto dphi_ci
+        = xt::view(dphi_c, i, xt::all(), xt::all(), xt::all(), xt::all());
+    dphi_ci = xt::view(cmap_basis_functions, xt::xrange(1, int(tdim) + 1),
+                       xt::all(), xt::all(), xt::all());
+  }
+
+  // Prepare output variables
+  std::vector<PetscScalar> circumradius;
+  circumradius.reserve(num_facets);
+  const int cstride = 1;
+
+  for (auto facet : active_facets)
+  {
+
+    // NOTE Add two separate loops here, one for and one without dof
+    // transforms
+
+    // FIXME: Assuming exterior facets
+    // get cell/local facet index
+    auto cells = f_to_c->links(facet);
+    // since the facet is on the boundary it should only link to one cell
+    assert(cells.size() == 1);
+    auto cell = cells[0]; // extract cell
+
+    // find local index of facet
+    auto cell_facets = c_to_f->links(cell);
+    auto local_facet = std::find(cell_facets.begin(), cell_facets.end(), facet);
+    const auto local_index = std::distance(cell_facets.data(), local_facet);
+
+    // Get cell geometry (coordinate dofs)
+    auto x_dofs = x_dofmap.links(cell);
+    for (std::size_t i = 0; i < num_dofs_g; ++i)
+    {
+      const int pos = 3 * x_dofs[i];
+      for (int j = 0; j < gdim; ++j)
+        coordinate_dofs(i, j) = x_g[pos + j];
+    }
+
+    // Compute determinant of Jacobian which is used to compute the area/volume
+    // of the cell
+    auto dphi_ci = xt::view(dphi_c, local_index, xt::all(), xt::all(),
+                            xt::all(), xt::all());
+    J.fill(0);
+    auto _J = xt::view(J, 0, xt::all(), xt::all());
+    xt::xtensor<double, 2> dphi
+        = xt::view(dphi_c, local_index, xt::all(), 0, xt::all(), 0);
+    dolfinx::fem::CoordinateElement::compute_jacobian(dphi, coordinate_dofs,
+                                                      _J);
+    detJ[0] = dolfinx::fem::CoordinateElement::compute_jacobian_determinant(_J);
+
+    // Sum up quadrature contributions
+    // NOTE: Consider refactoring (moving in Jacobian computation when we start
+    // supporting non-affine geoemtries)
+    circumradius.push_back(
+        compute_circumradius(mesh, detJ[0], coordinate_dofs));
+  }
+  return {std::move(circumradius), cstride};
 }
